+++ front-api/routes/poke/workspaces/[wId]/mcp/views/index.ts
@@ -1,14 +1,10 @@
-import type { MCPServerViewType } from "@app/lib/api/mcp";
+import type { PokeListMCPServerViews } from "@app/lib/api/poke/mcp_server_views";
 import { MCPServerViewResource } from "@app/lib/resources/mcp_server_view_resource";
 import { SpaceResource } from "@app/lib/resources/space_resource";
 import { pokeApp } from "@front-api/middlewares/ctx";
 import type { HandlerResult } from "@front-api/middlewares/utils";
 import { validate } from "@front-api/middlewares/validator";
 import { z } from "zod";
-
-export type PokeListMCPServerViews = {
-  serverViews: MCPServerViewType[];
-};
 
 const QuerySchema = z.object({
   globalSpaceOnly: z.enum(["true", "false"]).optional(),
