+++ front-api/routes/poke/workspaces/[wId]/mcp_server_views/[svId]/details.ts
@@ -1,14 +1,10 @@
+import type { PokeGetMCPServerViewDetails } from "@app/lib/api/poke/mcp_server_views";
 import { mcpServerViewToPokeJSON } from "@app/lib/poke/utils";
 import { MCPServerViewResource } from "@app/lib/resources/mcp_server_view_resource";
-import type { PokeMCPServerViewType } from "@app/types/poke";
 import { pokeApp } from "@front-api/middlewares/ctx";
 import { apiError, type HandlerResult } from "@front-api/middlewares/utils";
 import { validate } from "@front-api/middlewares/validator";
 import { z } from "zod";
-
-export type PokeGetMCPServerViewDetails = {
-  mcpServerView: PokeMCPServerViewType;
-};
 
 const ParamsSchema = z.object({
   svId: z.string(),
