+++ front-api/routes/poke/workspaces/[wId]/webhook_sources/index.ts
@@ -1,15 +1,9 @@
+import type { PokeListWebhookSources } from "@app/lib/api/poke/webhook_sources";
 import { listWebhookSourcesWithCounts } from "@app/lib/api/webhook_source";
-import type { WebhookSourceType } from "@app/types/triggers/webhooks";
 import { pokeApp } from "@front-api/middlewares/ctx";
 import type { HandlerResult } from "@front-api/middlewares/utils";
 
 import wsId from "./[wsId]";
-
-export type PokeListWebhookSources = {
-  webhookSources: Array<
-    WebhookSourceType & { viewCount: number; triggerCount: number }
-  >;
-};
 
 // Mounted at /api/poke/workspaces/:wId/webhook_sources.
 const app = pokeApp();
