+++ front/lib/api/poke/groups.ts
@@ -1,0 +1,11 @@
+import type { GroupType } from "@app/types/groups";
+import type { UserTypeWithWorkspaces } from "@app/types/user";
+
+export type PokeListGroups = {
+  groups: GroupType[];
+};
+
+export type PokeGetGroupDetails = {
+  members: UserTypeWithWorkspaces[];
+  group: GroupType;
+};
