+++ front/pages/api/poke/workspaces/[wId]/groups/index.ts
@@ -1,17 +1,13 @@
 /** @ignoreswagger */
 // @migration-status: MIGRATED_TO_HONO
 import { withSessionAuthenticationForPoke } from "@app/lib/api/auth_wrappers";
+import type { PokeListGroups } from "@app/lib/api/poke/groups";
 import { Authenticator } from "@app/lib/auth";
 import type { SessionWithUser } from "@app/lib/iam/provider";
 import { GroupResource } from "@app/lib/resources/group_resource";
 import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types/error";
-import type { GroupType } from "@app/types/groups";
 import type { NextApiRequest, NextApiResponse } from "next";
-
-export type PokeListGroups = {
-  groups: GroupType[];
-};
 
 async function handler(
   req: NextApiRequest,
