+++ front/pages/api/poke/workspaces/[wId]/mcp/views/index.ts
@@ -1,7 +1,7 @@
 /** @ignoreswagger */
 // @migration-status: MIGRATED_TO_HONO
 import { withSessionAuthenticationForPoke } from "@app/lib/api/auth_wrappers";
-import type { MCPServerViewType } from "@app/lib/api/mcp";
+import type { PokeListMCPServerViews } from "@app/lib/api/poke/mcp_server_views";
 import { Authenticator } from "@app/lib/auth";
 import type { SessionWithUser } from "@app/lib/iam/provider";
 import { MCPServerViewResource } from "@app/lib/resources/mcp_server_view_resource";
@@ -9,10 +9,6 @@ import { SpaceResource } from "@app/lib/resources/space_resource";
 import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types/error";
 import type { NextApiRequest, NextApiResponse } from "next";
-
-export type PokeListMCPServerViews = {
-  serverViews: MCPServerViewType[];
-};
 
 async function handler(
   req: NextApiRequest,
