+++ front/pages/api/poke/workspaces/[wId]/mcp_server_views/[svId]/details.ts
@@ -1,18 +1,14 @@
 /** @ignoreswagger */
 // @migration-status: MIGRATED_TO_HONO
 import { withSessionAuthenticationForPoke } from "@app/lib/api/auth_wrappers";
+import type { PokeGetMCPServerViewDetails } from "@app/lib/api/poke/mcp_server_views";
 import { Authenticator } from "@app/lib/auth";
 import type { SessionWithUser } from "@app/lib/iam/provider";
 import { mcpServerViewToPokeJSON } from "@app/lib/poke/utils";
 import { MCPServerViewResource } from "@app/lib/resources/mcp_server_view_resource";
 import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types/error";
-import type { PokeMCPServerViewType } from "@app/types/poke";
 import type { NextApiRequest, NextApiResponse } from "next";
-
-export type PokeGetMCPServerViewDetails = {
-  mcpServerView: PokeMCPServerViewType;
-};
 
 async function handler(
   req: NextApiRequest,
