+++ front/poke/swr/webhook_sources.ts
@@ -1,5 +1,5 @@
+import type { PokeListWebhookSources } from "@app/lib/api/poke/webhook_sources";
 import { useFetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
-import type { PokeListWebhookSources } from "@app/pages/api/poke/workspaces/[wId]/webhook_sources";
 import type { PokeConditionalFetchProps } from "@app/poke/swr/types";
 import type { Fetcher } from "swr";
